Create value_it.cpp
I am most greatly confused in how to access elements of the vectors and which are the ones I should be accessing. Is the probability a constant or is that part of the equation supposed to be a different vector? Also, should R(s) be a vector access or just the reward argument that is passed in. Let me know if and when you'd like to meet up.